+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixes
 
 ### Added
+- Show timestamp in occ commands [#88](https://github.com/rullzer/previewgenerator/pull/88)
 
 ## 1.0.8 - 2017-12-21